refactor booking form: extract date and time selectors into individual components
